refactor(Notification): Slack 도메인 패키지 위치 이동
